refactor: 참여자 연관 객체 패키지 분리
