+++ src/main/java/model/paticipant/Name.java
@@ -1,4 +1,4 @@
-package model;
+package model.paticipant;
 
 public record Name(String name) {
