+++ src/main/java/model/paticipant/Player.java
@@ -1,4 +1,4 @@
-package model;
+package model.paticipant;
 
 public class Player extends Participant {
