+++ src/test/java/model/DealerBustTest.java
@@ -6,6 +6,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import model.card.Card;
 import model.card.CardShape;
 import model.card.CardValue;
+import model.paticipant.Dealer;
 import org.junit.jupiter.api.Test;
 
 public class DealerBustTest {
