+++ src/test/java/model/JudgementTest.java
@@ -5,6 +5,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
 import model.card.Card;
+import model.paticipant.Dealer;
+import model.paticipant.Player;
+import model.paticipant.Players;
 import model.service.Judgement;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
