+++ src/test/java/model/PlayerBustTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import model.card.Card;
 import model.card.CardShape;
 import model.card.CardValue;
+import model.paticipant.Player;
 import org.junit.jupiter.api.Test;
 
 public class PlayerBustTest {
